reactivted poisson_3d_hex example: doesn't work though
the example runs through, but there is a pattern.
There are problems with non-simplex 2d and 3d domain configurations.
This might be related to the transform/dt_dx* implementations.

In the simplex case, numerical values (aka double) are stored on the
cells for the det_dF_dt and dt_dx_key<n,m> quantities. In this case it works,
however, with the non-simplex (cubic) based configurations, instead of numerical
values, viennamath::expr are stored. So maybe the handling of these expression-based
cell quantities in the assembler is flawed?!