fixed the hexa examples problem - all examples are ready including latex output
problem description:
It had infact something to do with the fact that the hexa-based transformations
use viennamath::expr instead of numerical values. The error was then located in
the cell_quan class, with the 'wrap_expr' method, where instead of a
viennamath::expr, a numeric_type was used for the value type. Of course this clashed then.
Adjusting the 'cell_quan_expr' class with respect to evaluating the 'accessed expression'
based on a given value, fixed the problem finally.